requirements.txt
onnxruntime>=1.18.0
Pillow>=9.0.0
numpy>=1.24.0
huggingface_hub>=0.23.0
